+++ package.json
@@ -29,7 +29,8 @@
     "prebuild": "node scripts/generate-sw-version.mjs",
     "build": "next build",
     "analyze": "ANALYZE=1 next build",
-    "bundle-report": "node scripts/print-bundle-report.mjs",
+    "bundle-report": "node scripts/check-bundle-budget.mjs",
+    "bundle-check": "node scripts/check-bundle-budget.mjs --check",
     "start": "next start",
     "lint": "eslint",
     "db:generate": "prisma generate",
