+++ src/app/__tests__/dashboard-suspense-boundaries.test.ts
@@ -23,7 +23,7 @@ import { join } from "node:path";
  * rather than silently regressing the streaming-composition contract.
  */
 const ROOT = join(__dirname, "../../..");
-const PAGE_PATH = join(ROOT, "src/app/page.tsx");
+const PAGE_PATH = join(ROOT, "src/app/page-client.tsx");
 
 function load(path: string): string {
   return readFileSync(path, "utf8");
